% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{airbnb}
\alias{airbnb}
\title{Racial Discrimination in the Sharing Economy}
\format{
a tibble containing:
\itemize{
\item acceptance: \code{yes} if the application is acccepted
\item guest_race: guest race (\code{black} or \code{other}
\item guest_gender: guest gender
\item host_race: host race (\code{black} or \code{other}
\item host_gender: host gender
\item multlistings: guest has several listings
\item shared: a dummy for shared properties
\item tenreviews: host has ten or more reviews
\item price: the price for a night
\item city: the city, one of \code{Balitmore}, \code{Dallas}, \code{Los-Angeles}, \code{St-Louis} or \code{Washington}
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 6235 applications from 2015
}
\references{
\insertRef{EDEL:LUCA:SVIR:17}{micsr}
}
\keyword{dataset}
