% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{apples}
\alias{apples}
\title{Apple production}
\format{
a tibble containing:
\itemize{
\item id: farm's id
\item year: year
\item capital: capital stock
\item labor: quantity of labor
\item materials: quantity of materials
\item apples: production of apples
\item otherprod: other productions
\item pc: price of capital
\item pl: price of labor
\item pm: price of materials
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
yearly observations of 173 farms from 1984 to 1986
}
\references{
\insertRef{IVAL:LADO:OSSA:SIMI:96}{micsr}
}
\keyword{dataset}
