% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{birthwt}
\alias{birthwt}
\title{Cigarette smoking and birth weight}
\format{
a tibble containing:
\itemize{
\item birthwt: birth weight
\item cigarettes: number of cigarettes smoked per day during pregnancy
\item parity: birth order
\item race: a factor with levels other and white
\item sex: a factor with levels female and male
\item edmother: number of years of education of the mother
\item edfather: number of years of education of the father
\item faminc: family income
\item cigtax: per-pack state excise tax on cigarettes
}
}
\source{
kindly provided by John Mullahy
}
\description{
a cross-section of 1388 individuals from 1988
}
\references{
\insertRef{MULL:97}{micsr}
}
\keyword{dataset}
