% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cartels}
\alias{cartels}
\title{Leniency and Cartels Enforcement}
\format{
a tibble containing:
\itemize{
\item semiyear: the semester
\item ncaught: the number of cartels discoveries
\item len: one if the period postdates august 10, 1993, new leniency program introduced by the Department Of Justice
\item time: a time trend, from 1 to 40
\item lent: a time trend starting one period after the introduction of the leniency program, zero before
\item dgdp: semestrial growth of the gdp
\item funds: average antitrust division budget allocation
\item fines: total corporate fines issued by the antitrust division during the previous fiscal year
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a time series of 40 semesterly observations from 1985 to 2005
}
\references{
\insertRef{MILL:09}{micsr}
}
\keyword{dataset}
