% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cigmales}
\alias{cigmales}
\title{Cigarette smoking behaviour}
\format{
a tibble containing:
\itemize{
\item cigarettes: number of daily cigarettes smoked
\item habit: smoking habit stock measure
\item price: state-level average per-pack price of cigarettes in 1979
\item restaurant: an indicator of whether the individual's state of residence had restrictions on smoking in restaurants in place in 1979
\item income: family income in thousands
\item age: age in years
\item educ: schooling in years
\item famsize: number of family members
\item race: a factor with levels "other" and "white"
\item reslgth: number of years the state's restaurant smoking restrictions had been in place in 1979
\item lagprice: one-year lag of cigarette price
}
}
\source{
kindly provided by John Mullahy
}
\description{
a cross-section of 6160 individuals from 1979 to 1980
}
\references{
\insertRef{MULL:97}{micsr}
}
\keyword{dataset}
