% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{coughlin}
\alias{coughlin}
\title{Populist Persuasion, Evidence from Father Coughlin}
\format{
a tibble containing:
\itemize{
\item state: state name
\item county: county name
\item id: county id
\item fdr: share of votes for Franklin D. Roosevelt in the 1936 presidential election
\item signal: signal strength of the Father Coughlin's radio program in 1936,
\item signal_free: hypothetical signal strenght (on an earth free of topographic obstacles)
\item male: share of males
\item native_whites: share of native whites
\item foreign_whites: share of foreign whites
\item blacks: share of blacks
\item urban: share of urbans
\item elderlies: share of the population aged 65 or more
\item catholics: share of catholics
\item illiterate: share of illeterates
\item unemp: unemployment rate
\item income: occupational income score
\item radio: share of radio owners
\item manuf: share of manufacturing workers
\item agri: share of agricultural workers
\item farm_size: average farm size
\item land_value: land value per acre
\item tenant_acre: share of tenant acres
\item area: area
\item elev: mean elevation
\item rug: rug
\item pop: population
\item past_democrat: average vote share of the democratic party
\item past_republican: average vote share of the republican party
\item past_turnout: aerage voter turnout during 1920-1928
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 3102 counties from 1936
}
\references{
\insertRef{WANG:21}{micsr}
}
\keyword{dataset}
