% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{doctor_aus}
\alias{doctor_aus}
\title{Doctor visits in Australia}
\format{
a tibble containing:
\itemize{
\item sex: a factor with levels \code{male} and \code{female}
\item age: age in years divided by 100
\item income: annual income in tens of thousands of dollars
\item insurance: insurance contract, a factor with levels \code{medlevy}, \code{levyplus} (private insurance), \code{freepor} (free government insurance due to low-income) and \code{freerepa} (free government insurance due to old-age)
\item illness: number of illness in past 2 weeks
\item actdays: number of days of reduced activity in past 2 weeks due to illness or injury
\item hscore: general health score using Goldberg's method (from 0 to 12)
\item chcond: chronic condition (\code{np} : no problem, \code{nla} : chronic condition but not limited in activity, \code{la} : chronic condition and milited in activity)
\item doctorco: number of consultations with a doctor or specialist in the past 2 weeks
\item nondocco: number of consultations with non-doctor health professionals (chemist, optician, physiotherapist, social worker, district community nurse, chiropodist or chiropractor) in the past 2 weeks
\item hospadmi: number of admissions to a hospital, psychiatric hospital, nursing or convalescent home in the past 12 months (up to 5 or more admissions which is coded as 5)
\item hospdays: number of nights in a hospital, etc.  during most recent admission: taken, where appropriate, as the mid-point of the intervals 1, 2, 3, 4, 5, 6, 7, 8-14, 15-30, 31-60, 61-79 with 80 or more admissions coded as 80. If no admission in past 12 months then equals zero.
\item medicine: total number of prescribed and nonprescribed medications used in past 2 days
\item prescrib: total number of prescribed medications used in past 2 days
\item nonpresc: total number of nonprescribed medications used in past 2 days
}
}
\source{
http://cameron.econ.ucdavis.edu/racd/racddata.html
}
\description{
a cross-section of 5190 individuals from 1977 to 1978
}
\references{
\insertRef{CAME:TRIV:86}{micsr}

\insertRef{CAME:JOHA:97}{micsr}

\insertRef{CAME:TRIV:MILN:PIGG:88}{micsr}

\insertRef{CAME:TRIV:13}{micsr}
}
\keyword{dataset}
