% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{drinks}
\alias{drinks}
\title{Physician advice on alcohol consumption}
\format{
a tibble containing:
\itemize{
\item drinks: number of drinks in the past 2 weeks
\item advice: 1 if reveived a drining advice
\item age: age in 10 years cathegories
\item race: a factor with levels \code{white}, \code{black} and \code{other}
\item marital: marital status, one of \code{single}, \code{married}, \code{widow}, \code{separated}
\item region: one of \code{west}, \code{northeast}, \code{midwest} and \code{south}
\item empstatus: one of \code{other}, \code{emp} and \code{unemp}
\item limits: limits on daily activities, one of \code{none}, \code{some} and \code{major}
\item income: monthly income ($1000)
\item educ: education in years
\item medicare: insurance through medicare
\item medicaid: insurance through medicaid
\item champus: military insurance
\item hlthins: health insurance
\item regmed: regoular source of care
\item dri: see same doctor
\item diabete: have diabetes
\item hearthcond: have heart condition
\item stroke: have stroke
}
}
\source{
JAE data archive
}
\description{
a cross-section of 2467 individuals from 1990
}
\references{
\insertRef{KENK:TERZ:01}{micsr}
}
\keyword{dataset}
