% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{federiv}
\alias{federiv}
\title{Foreign exchange derivatives use by large US bank holding companies}
\format{
a tibble containing:
\itemize{
\item federiv: foreign exchange derivatives use, a dummy
\item optval: option awards
\item eqrat: leverage
\item bonus: bonus
\item ltass: logarithm of total assets
\item linsown: logarithm of the percentage of the total shares outstanding that are owned by officers and directors
\item linstown: logarithm of the percentage of the total shares outstanding that are owned by all institutional investors
\item roe: return on equity
\item mktbk: market to book ratio
\item perfor: foreign to total interest income ratio
\item dealdum: derivative dealer activity dummy
\item div: dividends paid
\item year: year, from 1996 to 2000
\item no_emp: number of employees
\item no_subs: number of subsidiaries
\item no_off: number of offices
\item ceo_age: CEO age
\item gap: 12 month maturity mismatch
\item cfa: ratio of cash flow to total assets
}
}
\source{
Lee Adkin's home page https://learneconometrics.com/
}
\description{
a cross-section of 794 banks from 1996 to 2000
}
\references{
\insertRef{ADKI:12}{micsr}

\insertRef{ADKI:CART:SIMP:07}{micsr}
}
\keyword{dataset}
