% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{feesadm}
\alias{feesadm}
\title{Demand for fees and admission from the US expense survey}
\format{
a tibble containing:
\itemize{
\item feesadm: expense in fees and admission
\item linc: log of income per consumption unit
\item smsa: a dummy indicating if the household leaves in a MSA
\item size: number of members in the household
\item age: the age of the head of the household
\item educ: number of years of education of the head of the household
}
}
\source{
US bureau of labour statistics \url{https://www.bls.gov/cex/}
}
\description{
a cross-section of 6927 households from 2019
}
\keyword{dataset}
