% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{food}
\alias{food}
\title{Food consumption in the Netherlands}
\format{
a tibble containing:
\itemize{
\item year: year of the survey, either 1980 or 1988
\item weights: weight in the survey
\item hsize: number of persons in the household
\item ageh: age of the head of the household in 5 classes
\item income: total income
\item food: food expenditure
\item midage: a dummy for household for which the head is aged between 35 and 64 years old
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 4611 households
}
\examples{
tobit1(log(food) ~ log(income) + log(hsize) + midage,
       data = food, subset = year == 1980,
       left = -Inf, right = log(13030))
}
\references{
\insertRef{CROM:PALM:URBA:97}{micsr}
}
\keyword{dataset}
