% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{ftest}
\alias{ftest}
\alias{ftest.lm}
\alias{ftest.ivreg}
\title{F statistic}
\usage{
ftest(x, ...)

\method{ftest}{lm}(x, ...)

\method{ftest}{ivreg}(x, ..., covariate = NULL)
}
\arguments{
\item{x}{a fitted object}

\item{...}{further arguments}

\item{covariate}{the covariate for which the test should be performed for the \code{ivreg} method}
}
\value{
an object of class \code{"htest"}.
}
\description{
Extract the F statistic that all the parameters except the
intercept are zero. Currently implemented only for models fitted by \code{lm} or \code{ivreg::ivreg}.
}
\keyword{htest}
