% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gauss_laguerre}
\alias{gauss_laguerre}
\title{Gauss-Laguerre quadrature}
\usage{
gauss_laguerre(N)
}
\arguments{
\item{N}{the number of evaluations}
}
\value{
a list containing two numeric vectors of length N, the first one containing the nodes and the second one the weights
}
\description{
Computes the node and the weights for the Gauss-Laguerre quadrature (integral on the whole real line)
}
