% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{growth}
\alias{growth}
\title{Augmented Sollow's growth model with human capital}
\format{
a tibble containing:
\itemize{
\item country: country
\item group: country's group, a factor with levels \code{oecd}, \code{other}, lqdata\verb{(for low quality data) and}oil` (oil exporter countries)
\item gdp60: per-capita gdp in 1960
\item gdp85: per-capita gdp in 1985
\item gdpgwth: mean annual growth rate of the gdp
\item popgwth: mean annual growth rate of the population
\item inv: mean value of the investment rate
\item school: measurment of human capital
\item growth: log-difference of per-capita gdp between 1985 and 1960
}
}
\source{
https://github.com/HariharanJayashankar/mrw1992
}
\description{
a cross-section of 121 countries from 1985
}
\references{
\insertRef{MANK:ROME:WEIL:92}{micsr}
}
\keyword{dataset}
