% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{hausman}
\alias{hausman}
\alias{hausman.ivreg}
\alias{hausman.micsr}
\title{Hausman test}
\usage{
hausman(x, y, omit = FALSE, ...)

\method{hausman}{ivreg}(x, y, omit = FALSE, ...)

\method{hausman}{micsr}(x, y, omit = NULL, ...)
}
\arguments{
\item{x}{the first model,}

\item{y}{the second model}

\item{omit}{a character containing the effects that are removed from the test}

\item{...}{further arguments}
}
\value{
an object of class \code{"htest"}.
}
\description{
Hausman test; under the null both models are consistent but one of
them is more efficient, under the alternative, only one model is
consistent
}
\references{
\insertRef{HAUS:78}{micsr}
}
\author{
Yves Croissant
}
\keyword{htest}
