% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{housprod}
\alias{housprod}
\title{Household Production}
\format{
a tibble containing:
\itemize{
\item mjob: dummy, 1 if male has paid job
\item fjob: dummy, 1 if female has paid job
\item mtime: home production time male (minutes per day)
\item ftime: home production time female (minutes per day)
\item mwage: net hourly wage rate male (estimate imputed if mjob=0)
\item fwage: net hourly wage rate female (estimate imputed if fjob=0)
\item mage: age male
\item meduc: years of schooling male
\item fage: age female
\item feduc: years of schooling female
\item owner: dummy, 1 if houseownwers
\item fsize: family size
\item ychild: number of children younger than 7 years old in the household
\item cars: number of cars in the household
\item nonlabinc: non-labour income (in units of 1000 Swedish Kronor)
}
}
\source{
JAE data archive
}
\description{
a cross-section of 819 households from 1984
}
\references{
\insertRef{KERK:KOOR:03}{micsr}
}
\keyword{dataset}
