% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{lakisha}
\alias{lakisha}
\title{Resumes with 'white' or 'african-american' names}
\format{
a tibble containing:
\itemize{
\item call: has the employer responded to the resume ?
\item firstname: a factor of 36 first names (18 for males and 18 for females, 18 for blacks and 18 for whites)
\item city: one of 'boston' and 'chicago'
\item sex: one of 'female', 'male'
\item race: one of 'black' and 'white'
\item quality: a subjective measure of the quality of the resume, one of 'low' and 'high'
\item exp: years of experience
\item volunteer: volunteering experience ?
\item military: military experience ?
\item email: email address ?
\item empholes: employment holes ?
\item worksch: work in school ?
\item honors: honors ?
\item compsk: computer skills ?
\item specsk: special skills ?
\item apblack: fraction african-americans in applicant's zip code
\item apwhite: fraction whites in applicant's zip code
\item apdrop: fraction high school dropouts in applicant's zip code
\item apcolp: fraction college or more in applicant's zip code
\item alinc: log of median household income in applicant's zip code
\item occupation: the occupation, a factor with 6 levels
\item sector: the secor, a factor with 7 levels
\item req: any requirement ?
\item reqexp: experience requirement ?
\item reqcom: communication skills required ?
\item reqeduc: education requirement ?
\item reqcomp: computer skills required ?
\item reqorg: organization skills ?
\item fed: federal contractor ?
\item eoe: equal opportunity employer ?
\item ownership: ownership status ?
\item epblack: fraction african-americans in employer's zip code
\item epwhite: fraction whites in employer's zip code
\item epdrop: fraction high school dropouts in employer's zip code
\item epcolp: fraction college or more in employer's zip code
\item elinc: log of median household income in employer's zip code
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 4870 resumes from 2001 to 2002
}
\references{
\insertRef{BERT:MULL:04}{micsr}
}
\keyword{dataset}
