% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{loan_market}
\alias{loan_market}
\title{US commercial loan market}
\format{
a tibble containing:
\itemize{
\item date: the month of observation
\item loans: total commercial loans, billions of dollars
\item prime_rate: average prime rate charged by banks
\item aaa_rate: AAA corporate bond rate (alternative financing for firms)
\item ipi: industrial production index (firms' expectation about future economic activity)
\item treas_rate: 3-month treasury bill rate (alternative rate of return for banks)
\item deposits: total bank deposit (scale variable), billions of dollars
}
}
\source{
G.S. Maddala thanks Water Mayer for providing him with the data (p.364-65)
}
\description{
a time series of 72 monthly observations from 1979-01 to 1984-12
}
\references{
\insertRef{MADD:01}{micsr}
}
\keyword{dataset}
