% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglm.R
\name{loglm}
\alias{loglm}
\title{Log-linear model}
\usage{
loglm(formula, data)
}
\arguments{
\item{formula, data}{see \code{lm}}
}
\value{
An object of class \code{micsr}
}
\description{
Estimation of log-linear model; the estimation is done by \code{lm}, but
the correct log-likelihood related quantities are returned
}
\examples{
lm(dist ~ log(speed), cars)
loglm(dist ~ log(speed), cars)
}
\author{
Yves Croissant
}
