% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{maximize}
\alias{maximize}
\title{Maximization of a function}
\usage{
maximize(
  x,
  start,
  method = c("bfgs", "nr", "newton"),
  trace = 0,
  maxit = 100,
  ...
)
}
\arguments{
\item{x}{the function to maximize}

\item{start}{a vector of starting values}

\item{method}{the optimization method}

\item{trace}{if positive or true, some information about the
computation is printed}

\item{maxit}{maximum number of iterations}

\item{...}{further arguments, passed to the function}
}
\value{
a numeric vector, the parameters at the optimum of the
function.
}
\description{
This function provides a unified interface to three optimization
algorithms: the BFGS algorithm provided by \code{stats::optim}, the
Newton-Ralphson algorithm provided by \code{stats::nlm} and a simple
Newton-Ralphson algorithm provided by \code{micsr::newton}
}
\keyword{misc}
