% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mills}
\alias{mills}
\title{Compute the inverse Mills ratio and its first two derivatives}
\usage{
mills(x, deriv = 0)
}
\arguments{
\item{x}{a numeric}

\item{deriv}{one of 0 (the default, returns the inverse Mills
ratio), 1 (the first derivative) and 2 (the second derivative)}
}
\value{
a numeric
}
\description{
The inverse Mills ratio is used in several econometric models,
especially different flavours of tobit model.
}
\keyword{misc}
