% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mode_choice}
\alias{mode_choice}
\title{Choice between car and transit}
\format{
a tibble containing:
\itemize{
\item mode: 1 for car, 0 for transit
\item cost: transit fare minus automobile travel cost in US$
\item ivtime: transit in-vehicule travel time minus in-vehicule travel time (minutes)
\item ovtime: transit out-of vehicule time minus out-of vehicule travel time (minutes)
\item cars: number of cars owned by the traveler's household
}
}
\source{
GAMS's website \url{http://www.gams.com/modlib/libhtml/mws.htm}
}
\description{
a cross-section of 842 individuals
}
\references{
\insertRef{HORO:93}{micsr}
}
\keyword{dataset}
