% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbnorm}
\alias{pbnorm}
\title{Compute the probability for the bivariate normal function}
\usage{
pbnorm(z1, z2, rho)
}
\arguments{
\item{z1, z2}{two numeric vectors}

\item{rho}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Compute the probability for the bivariate normal function
}
