% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{price_time}
\alias{price_time}
\title{Market shares of train and air for Paris and main French cities in 1995}
\format{
a tibble containing:
\itemize{
\item town: the city
\item region: one of \code{SW}, \code{SE} and \code{O}
\item trafic_rail: number of rail passengers
\item trafic_air: number of air passengers
\item price_rail: price of a train ticket in euros
\item price_air: price of a plane ticket in euros
\item time_rail: length of a train trip in minutes
\item time_air: length a plane trip in minutes
}
}
\source{
Patrick Bonnel's book
}
\description{
a cross-section of 13 cities from 1995
}
\references{
\insertRef{BONN:04}{micsr}
}
\keyword{dataset}
