% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{private_ins}
\alias{private_ins}
\title{Private insurance and doctor visits}
\format{
a tibble containing:
\itemize{
\item size: family size
\item smsa: lives in a msa
\item age: age
\item sex: sex
\item educ: number of years of education
\item wage: wage
\item nbemp: number of employees in the firm
\item pluriloc: a dummy for firms with multiple locations
\item privateins: a dummy for private insurance
\item doctor: a dummy if doctor visits
\item married: a dummy equal 1 if ever married
\item income: income
\item region: a factor for the region, one of \code{northeast}, \code{midwest}, \code{south} and \code{west}
\item race: a factor for the race, one of \code{white}, \code{black}, \code{minority} and \code{asian}
\item mental: a factor for mental health, one of \code{poor}, \code{fair}, \code{good}, \code{verygood} and \code{excellent}
\item physical: a factor for physical health, one of \code{poor}, \code{fair}, \code{good}, \code{verygood} and \code{excellent}
}
}
\source{
JAE data archive
}
\description{
a cross-section of 7555 individuals from 2010
}
\references{
\insertRef{HAN:LEE:19}{micsr}
}
\keyword{dataset}
