% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{prod_canada}
\alias{prod_canada}
\title{Canadian Manufacturing}
\format{
a tibble containing:
\itemize{
\item year: year
\item po: output price
\item qo: output volume
\item pm: price index for materials
\item qm: quantity of materials
\item pl: price index of productive labor
\item ql: quantity of productive labor
\item pn: price index of non-productive labor
\item qn: quantity of non-productive labor
\item k: capital stock
}
}
\source{
original paper, p.375
}
\description{
a time series of 25 yearly observations from 1946 to 1970
}
\references{
\insertRef{WOOD:77}{micsr}
}
\keyword{dataset}
