% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ptnorm}
\alias{ptnorm}
\title{Compute the probability for the trivariate normal function}
\usage{
ptnorm(z, rho)
}
\arguments{
\item{z}{a matrix with three columns}

\item{rho}{a matrix with three columns}
}
\value{
a numeric vector
}
\description{
Compute the probability for the trivariate normal function
}
