% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{punorm}
\alias{punorm}
\title{Compute the probability for the univariate normal function}
\usage{
punorm(z)
}
\arguments{
\item{z}{a numeric vector}
}
\value{
a numeric vector
}
\description{
Compute the probability for the univariate normal function
}
