% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{quad_form}
\alias{quad_form}
\title{Compute quadratic form}
\usage{
quad_form(x, m = NULL, inv = TRUE, subset = NULL, vcov = NULL, ...)
}
\arguments{
\item{x}{a numeric vector or a fitted model}

\item{m}{a square numeric matrix}

\item{inv}{a boolean, if \code{TRUE} (the default), the quadratic form
is computed using the inverse of the matrix}

\item{subset}{a subset of the vector and the corresponding subset
of the matrix}

\item{vcov}{if \code{NULL} the \code{vcov} method is used, otherwise it can
be a function or, for \code{micsr} objects, a character}

\item{\dots}{arguments passed to \code{vcov} if it is a function}
}
\description{
Compute quadratic form of a vector with a matrix, which can be the
vector of coefficients and the covariance matrix extracted from a
fitted model
}
