% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{random_group}
\alias{random_group}
\title{Random control group}
\format{
a tibble containing:
\itemize{
\item female: 1 for females
\item age: age
\item child: children
\item migrant: non-dutch
\item single: 1 for singles
\item temp: one for temporary job
\item ten: firm tenure (months)
\item edu: education, one of \code{"Low"}, \code{"Intermediate"} and \code{"High"}
\item fsize: firm size, one of \code{"up to 50"}, \code{"50 to 200"} and \code{"more than 200"}
\item samplew: sample weights
\item lnwh: log of hearly wage
\item group: group indicator, from -2 to 3
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 2166 individuals from 2001
}
\references{
\insertRef{LEUV:OOST:08}{micsr}
}
\keyword{dataset}
