% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{recall}
\alias{recall}
\title{recall}
\format{
a tibble containing:
\itemize{
\item id: individual id
\item spell: spell id
\item end: the situation at the end of the observation of the spell; a factor with levels \code{"new-job"}, \code{"recall"} or \code{"censored"}
\item duration: duration of unemployment spell
\item age: age the year before the spell
\item sex: a factor with levels \code{"male"} and \code{"female"}
\item educ: years of schooling
\item race: a factor with levels \code{"white"} and \code{"nonwhite"}
\item nb: number of dependents
\item ui: a factor indicating unemployment insurance during the spell
\item marital: marital status, a factor with levels \code{"single"} and \code{"married"}
\item unemp: county unemployment rate (interval midpoints for 1980 spells)
\item wifemp: wife's employment status, a factor with levels \code{"no"} and \code{"yes"},
\item homeowner: home owner, a factor with levels \code{"no"} and \code{"yes"},
\item occupation: a factor with 5 levels
\item industry: a factor with 9 levels
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 1045 spell of unemployment from 1980
}
\references{
\insertRef{SUEY:95}{micsr}
}
\keyword{dataset}
