% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/ndvuong.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Surv}
\alias{glance}
\alias{tidy}
\alias{bread}
\alias{estfun}
\alias{vcovHC}
\alias{model.part}
\alias{meat}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Formula}{\code{\link[Formula:model.part]{model.part()}}}

  \item{generics}{\code{\link[generics:glance]{glance()}}, \code{\link[generics:tidy]{tidy()}}}

  \item{sandwich}{\code{\link[sandwich:bread]{bread()}}, \code{\link[sandwich:bread]{bread()}}, \code{\link[sandwich:estfun]{estfun()}}, \code{\link[sandwich:estfun]{estfun()}}, \code{\link[sandwich:meat]{meat()}}, \code{\link[sandwich:vcovHC]{vcovHC()}}}

  \item{survival}{\code{\link[survival:Surv]{Surv()}}}
}}

