% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrreg.R
\name{rrreg}
\alias{rrreg}
\title{Recenter and rescale regression}
\usage{
rrreg(
  formula,
  data,
  subset = NULL,
  weights = NULL,
  na.action,
  offset,
  method = c("ols", "iv", "gmm"),
  model = c("probit", "tobit"),
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model,}

\item{data}{a data frame,}

\item{subset, weights, na.action, offset}{see \code{lm},}

\item{method}{the estimation method, one of \code{"ols"}, \code{"iv"} and \code{"gmm"}}

\item{model}{the model to estimate, one of \code{"probit"} and \code{"tobit"}}

\item{...}{further arguments}
}
\value{
an object of class \code{micsr}
}
\description{
A method of estimation for probit and tobit models which consist on
recentering and rescaling the response so that linear estimators can then be applied
}
\examples{
library("dplyr")
data("PSID1976", package = "AER")
Mroz <- PSID1976 \%>\% as_tibble \%>\%
      mutate(participation = as.numeric(participation == "yes"),
             nwincome = (fincome - hours * wage) / 1000, exper2 = I(experience ^ 2))
rrols <- rrreg(participation ~ age + education + youngkids + oldkids | . - education + meducation + feducation, data = Mroz)
rriv <- update(rrols, method = "iv")
rrgmm <- update(rrols, method = "gmm")
# Adkins (2012) : Bank
rrols <- rrreg(federiv ~ eqrat + optval + bonus + ltass + linsown + linstown +
                   roe + mktbk + perfor + dealdum + div + year |
                   . - eqrat - bonus - optval +
                   no_emp + no_subs + no_off + ceo_age + gap + cfa,
               data = federiv, method = "ols")
rriv <- update(rrols, method = "iv")
rrgmm <- update(rriv, method = "gmm")
}
