% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expreg.R
\name{sargantest}
\alias{sargantest}
\title{Sargan test for GMM models}
\usage{
sargantest(object, ...)
}
\arguments{
\item{object}{a model fitted by GMM}

\item{...}{further arguments}
}
\value{
an \code{htest} object
}
\description{
When a GMM model is over-identified, the set of all the empirical
moment conditions can't be exactly 0. The test of the validity of
the instruments is based on a quadratic form of the vector of the
empirical moments
}
