% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{scoretest}
\alias{scoretest}
\alias{scoretest.micsr}
\title{Score test}
\usage{
scoretest(x, y, ...)

\method{scoretest}{micsr}(x, y, ..., vcov = NULL)
}
\arguments{
\item{x}{the first model,}

\item{y}{the second model}

\item{...}{further arguments}

\item{vcov}{omit a character containing the effects that are
removed from the test}
}
\value{
a list with class \code{'htest'} containing the following
components: - data.mane: a character string describing the
fitted model - statistic: the value of the test statistic -
parameter: degrees of freedom - p.value: the p.value of the
test - method: a character indicating what type of test is
performed - alternative: a character indicating the alternative
hypothesis
}
\description{
Score test, also knowned as Lagrange multiplier tests
}
\author{
Yves Croissant
}
\keyword{htest}
