% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sibling_educ}
\alias{sibling_educ}
\title{Education of oldest sibling as an instrument of own education}
\format{
a tibble containing:
\itemize{
\item wage: hourly wage
\item educ: Education (years of schooling)
\item ability: ability based on standardized AFQT test score
\item educsib: education of oldest sibling (years of schooling)
\item experience: experience (total weeks of labor market experience)
\item tenure: tenure (weeks on current job)
\item mothed: Mother's education (years of schooling)
\item fathed: Father's education (years of schooling)
\item residence: one of urban or rural
\item broken: broken home dummy (dummy for living with both parents at age 14)
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 303 individuals from 1992
}
\references{
\insertRef{KOOP:POIR:TOBI:05}{micsr}
}
\keyword{dataset}
