% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sight.R
\name{sight}
\alias{sight}
\alias{sight.lm}
\alias{sight.micsr}
\alias{sight.ivreg}
\alias{sight.rdrobust}
\alias{sight.CJMrddensity}
\alias{sight.htest}
\alias{sight.anova}
\alias{sight.LMtestlist}
\title{Short print of the summary of an object}
\usage{
sight(x, ...)

\method{sight}{lm}(
  x,
  ...,
  coef = NULL,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = FALSE
)

\method{sight}{micsr}(
  x,
  ...,
  coef = NULL,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = FALSE
)

\method{sight}{ivreg}(
  x,
  ...,
  coef,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars")
)

\method{sight}{rdrobust}(x, ..., first_stage = FALSE)

\method{sight}{CJMrddensity}(x, ...)

\method{sight}{htest}(x, ..., digits = 3)

\method{sight}{anova}(x, ..., digits = 3)

\method{sight}{LMtestlist}(x, ..., digits = 3)
}
\arguments{
\item{x}{an object,}

\item{...}{further arguments for the different methods,}

\item{coef}{the coefficients to be printed}

\item{digits}{the number of digits for the \code{lm} and the \code{ivreg}
methods}

\item{signif.stars}{a boolean indicating whether the stars should
be printed}

\item{first_stage}{a boolean for the \code{rdrobust::rdrobust} method,
if \code{TRUE} the results of the first stage estimation,}
}
\value{
invisibly its first argument
}
\description{
\code{print} and \code{print.summary} methods often returns long input, which
is suitable for the console, but to verbal for a printed output
like a book or an article written using quarto. \code{sight} is a generic
function which prints a short output
}
