% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soep}
\alias{soep}
\title{Number of direct job changes from the German Socio-economic panel survey}
\format{
a tibble containing:
\itemize{
\item djc: number of direct job changes, i.e. changes without an intervening spell of unemployment,
\item lfp: a dummy for labor force participation,
\item exper: full-time employment in years
\item single: a factor indicating if the individual is single
\item whitecol: a foctor indicating if the individual is white collar
\item educ: lenght of education in years
\item spd: a factor indicating if the individual is a strong or a very strong supporter of the social-democrat party of Germany
}
}
\source{
this data is part of \code{SemiParSampleSel} package
}
\description{
a cross-section of 2651 individuals from 1984
}
\references{
\insertRef{WYSZ:MARR:18}{micsr}
}
\keyword{dataset}
