% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{stder}
\alias{stder}
\alias{stder.default}
\title{Extract the standard errors of estimated coefficients}
\usage{
stder(x, .vcov, ...)

\method{stder}{default}(x, .vcov = NULL, ...)
}
\arguments{
\item{x}{a fitted model or a matrix of covariance}

\item{.vcov}{a function that computes a covariance matrix, or a character}

\item{...}{further arguments}
}
\value{
a numeric vector
}
\description{
The standard errors are a key element while presenting the results
of a model. They are the second column of the table of coefficient
and are used to compute the t/z-value. \code{stderr} enables to retrieve
easily the vector of standard errors, either from a fitted model or
from a matrix of covariance
}
\keyword{misc}
