% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tracks_side}
\alias{tracks_side}
\title{Railroad tracks and segregation}
\format{
a tibble containing:
\itemize{
\item city: city
\item state: state
\item segregation: dissimilarity index in 1990 of the city
\item trackdens: track length per square kilometer
\item raildiv: railroad division index
\item giniw: gini index for whites
\item ginib: gini index for blacks
\item povw: poverty rate for whites
\item povb: poverty rate for blacks
\item D9wD9b: ratio of the ninth decile for whites and blacks
\item D1wD1b: ratio of the first decile for whites and blacks
\item D9wD1b: ratio of the ninth decile for whites and the first decile for blacks
\item D9bD1w: ratio of the ninth decile for blacks and the first decile for whites
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 121 cities from 1990
}
\references{
\insertRef{ANAN:11}{micsr}
}
\keyword{dataset}
