% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trade_protection}
\alias{trade_protection}
\title{Lobying from Capitalists and Unions and Trade Protection}
\format{
a tibble containing:
\itemize{
\item ntb: nontariff barrier coverage ratio
\item vshipped: value of shipments
\item imports: importations
\item elast: demand elasticity
\item cap: lobying
\item labvar: labor market covariate
\item sic3: 3-digit SIC industry classification
\item k_serv: physical capital, factor share
\item inv: Inventories, factor share
\item engsci: engineers and scientists, factor share
\item whitecol: white collar, factor share
\item skill: skilled, factor share
\item semskill: semi-skilled, factor share
\item cropland: cropland, factor shaer
\item pasture: pasture, factor share
\item forest: forest, factor share
\item coal: coal, factor share
\item petro: petroleum, factor share
\item minerals: minerals, factor share
\item scrconc: seller concentration
\item bcrconc: buyer concentration
\item scrcomp: seller number of firms
\item bcrcomp: buyer number of firms
\item meps: scale
\item kstock: capital stock
\item puni: proportion of workers union
\item geog2: geographic concentration
\item tenure: average worker tenure, years
\item klratio: capital-labor ratio
\item bunion:
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 194 United States
}
\references{
\insertRef{MATS:SHER:06}{micsr}
}
\keyword{dataset}
