% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{traffic_citations}
\alias{traffic_citations}
\title{Political Economy and Traffic Citations}
\format{
a tibble containing:
\itemize{
\item fine: \code{yes} for a fine, \code{no} for a warning,
\item amount: the amount of the fine if issued
\item black: dummy for a black offender,
\item hispanic: dummy for an hispanic offender,
\item age: the age of the offender
\item female: dummy for a female offender,
\item cdl: dummy for a commercial driving license,
\item res: the residence of the offender, one of \code{loc}, \code{oto} and \code{ost}, whether the offender resides in the town, out of the town in Massaschusetts or in another state
\item courtdist: distance from the distance courthouse
\item id: the id of the police officer
\item stpol: dummy for state officers
\item mph: the difference between actual and limit speed,
\item prval: the average property value of the municipality
\item hospemp: the percentage of employement in the hospitality sector
\item oloss: a dummy that indicates that the municipality failed to pass an overide referendum
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 68357 fines from 2001
}
\examples{
sel <- fine ~ log(mph) + black + hispanic + female * log(age) +
       stpol * (res + log(prval) + oloss) + cdl
out <- update(sel, log(amount) ~ . - cdl)
probit <- glm(sel, data = traffic_citations,
             family = binomial(link='probit'))
lp <- probit$linear.predictor
mls <- dnorm(lp) / pnorm(lp)
summary(probit)
# Computing the heckit by hand
lm <- lm(out, traffic_citations)
heck <- update(lm, . ~ . + mls)
# or using the heckit function
library("sampleSelection")
heck <- heckit(sel, out, data = traffic_citations,
            method = "2step")
summary(heck)
}
\references{
\insertRef{MAKO:STRA:09}{micsr}
}
\keyword{dataset}
