% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trips}
\alias{trips}
\title{Determinants of household trip taking}
\format{
a tibble containing:
\itemize{
\item trips: number of trips taken by a member of a household the day prior the survey interview
\item car: 1 if household owns at least one motorized vehicule
\item workschl: share of trips for work or school vs personal business or pleasure
\item size: number of individuals in the household
\item dist: distance to central business district in kilometers
\item smsa: a factor with levels \code{"small"} (less than 2.5 million population) and \code{"large"} (more than 2.5 million population)
\item fulltime: number of fulltime workers in household
\item adults: number of adults in household
\item distnod: distace from home to nearest transit node, in blocks
\item realinc: household income divided by median income of census tract in which household resides
\item weekend: 1 if the survey period is either saturday or sunday
}
}
\source{
kindly provided by Joseph Terza
}
\description{
a cross-section of 577 households from 1978
}
\references{
\insertRef{TERZ:98}{micsr}

\insertRef{TERZ:WILS:90}{micsr}
}
\keyword{dataset}
