% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{twa}
\alias{twa}
\title{Temporary help jobs and permanent employment}
\format{
a tibble containing:
\itemize{
\item id: identification code
\item age: age
\item sex: a factor with levels \code{"female"} and \code{"male"}
\item marital: marital status, \code{"married"} or \code{"single"}
\item children: number of children
\item feduc: father's education
\item fbluecol: father blue-color
\item femp: father employed at time 1
\item educ: years of education
\item pvoto: mark in last degree as fraction of max mark
\item training: received professional training before treatment
\item dist: distance from nearest agency
\item nyu: fraction of school-to-work without employment
\item hour: weekly hours of work
\item wage: monthly wage
\item hwage: hourly wage at time 1
\item contact: contacted a temporary work agency
\item region: one of  \code{"Tuscany"} and \code{"Sicily"}
\item city: the city
\item group: one of \code{"control"} and `"treated"
\item sector: the sector
\item occup: occupation, one of \code{"nojob"}, \code{"selfemp"}, \code{"bluecol"} and \code{"whitecol"}
\item empstat: employment status, one of \code{"empl"}, \code{"unemp"} and \code{"olf"} (out of labor force)
\item contract: job contract, one of \code{"nojob"}, \code{"atyp"} (atypical) and \code{"perm"} (permanent)
\item loc: localisation, one of \code{"nord"}, \code{"centro"}, \code{"sud"} and \code{"estero"}
\item outcome: one of \code{"none"}, \code{"other"}, \code{"fterm"} and \code{"perm"}
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 2030 individuals
}
\references{
\insertRef{ICHI:MEAL:NANN:08}{micsr}
}
\keyword{dataset}
