% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{twins}
\alias{twins}
\title{Education and earnings of british twins}
\format{
a tibble containing:
\itemize{
\item family: family index
\item earning: hourly wage rate
\item educ: reported years of schooling
\item educt: estimated years of schooling
\item age: age
\item londse: lives in
\item part: part-time job ?
\item married: married ?
\item tenure: tenure in years
}
}
\source{
American Economic Review data archive : \url{http://aeaweb.org/aer/}
}
\description{
a pseudo-panel of 214 families from 1999
}
\references{
\insertRef{BONJ:CHER:HASK:HAWK:SPEC:03}{micsr}
}
\keyword{dataset}
