% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UnempDuration}
\alias{UnempDuration}
\title{Unemployment Duration in Germany}
\format{
a tibble containing:
\itemize{
\item duration: the duration of the unemployment spell in days
\item censored: a factor with levels \code{yes} if the spell is censored, \code{no} otherwise
\item gender: a factor with levels \code{male} and \code{female}
\item age: the age
\item wage: the last daily wage before unemployment
}
}
\source{
The Royal Statistical Society Datasets Website \url{http://www.blackwellpublishing.com/rss/}
}
\description{
a cross-section of 21685 individuals from 1996 to 1997
}
\references{
\insertRef{WICH:WILK:08}{micsr}
}
\keyword{dataset}
