% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{urban_gradient}
\alias{urban_gradient}
\title{Density and distance to the CBD in Alabama}
\format{
a tibble containing:
\itemize{
\item msa: msa name
\item county: county name
\item tract: tract id
\item blkg: blkg id
\item area: area in square km
\item population: number of inhabitants
\item distance: distance to the CBD in km
}
}
\source{
American Economic Association Data Archive : \url{https://www.aeaweb.org/aer/}
}
\description{
a cross-section of 2315 blocks from 1999
}
\references{
\insertRef{DURA:PUGA:20}{micsr}
}
\keyword{dataset}
