% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vn_income}
\alias{vn_income}
\title{Draft eligibility and income}
\format{
a tibble containing:
\itemize{
\item birth: year of birth
\item year: year of observation for the income
\item interval: 5 days interval of date of birth (from 1 to 73)
\item type: income measurment, one of \code{fica} and \code{w2}
\item race: a factor with levels \code{white} and \code{nonwhite}
\item freq: the number of individuals in the cell
\item null_inc: share of null incomes in the cell
\item mean_inc: mean income in the cell
\item sd_inc: standard-deviation of income in the cell
}
}
\source{
Joshua Angrist's web page \href{https://economics.mit.edu/faculty/angrist}{}
}
\description{
a cross-section of 40880 cells of individuals
}
\references{
\insertRef{ANGR:90}{micsr}
}
\keyword{dataset}
