% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vn_veteran}
\alias{vn_veteran}
\title{Draft eligibility and enrolment in the Viet-Nam war}
\format{
a tibble containing:
\itemize{
\item birth: year of birth
\item wgts: weights
\item race: a factor with levels \code{white} and \code{nonwhite}
\item veteran: a factor with levels \code{yes} for veterans and \code{no} otherwise
\item eligible: a factor with levels \code{yes} for draft eligible men and \code{no} otherwise
}
}
\source{
Joshua Angrist's web page \href{https://economics.mit.edu/faculty/angrist}{}
}
\description{
a cross-section of 4545 individuals
}
\references{
\insertRef{ANGR:90}{micsr}
}
\keyword{dataset}
