% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wine}
\alias{wine}
\title{Price of wine}
\format{
a tibble containing:
\itemize{
\item price: wine price (cpi adjusted)
\item cases: cases produced
\item score: wsm tasting score
\item age: years of aging
\item region: region of production
\item variety: grape variety used
\item vintage: vintage, from 1991 to 1999
\item reserve: \code{'yes'} if reserve wine
\item vineyard: \code{'yes'} if vineyard information is provided
\item estate: \code{'yes'} if estate produced
\item class: cluster 1,2,3,4, per LPRC
}
}
\source{
Journal of Applied Econometrics Data Archive : \url{http://qed.econ.queensu.ca/jae/}
}
\description{
a cross-section of 9600 wines
}
\references{
\insertRef{COST:MITT:MCCL:09}{micsr}
}
\keyword{dataset}
