% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmeans.R
\docType{data}
\name{binmeans}
\alias{binmeans}
\alias{binmeans.default}
\alias{binmeans.formula}
\alias{StatBinmeans}
\alias{geom_binmeans}
\title{Compute the mean of a variable for bins of another variable}
\format{
An object of class \code{StatBinmeans} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
binmeans(x, y, width = NULL, center = NULL, g = NULL, ...)

\method{binmeans}{default}(x, y, width = NULL, center = NULL, g = NULL, ..., name_g = "colour")

\method{binmeans}{formula}(x, y, width = NULL, center = NULL, ...)

StatBinmeans

geom_binmeans(
  mapping = NULL,
  data = NULL,
  stat = "binmeans",
  position = "identity",
  ...,
  center = NULL,
  width = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{x, y}{either two numeric vector for the default method, or a
formula and a data frame for the formula method}

\item{width}{the width of the bins}

\item{center}{the cuting value of the forcing variable}

\item{g}{a grouping variable}

\item{...}{further arguments}

\item{name_g}{internally used by the geom}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used the override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Plot average values of the outcome for bins of the forcing
variable, a common plot in regression discontinuity analysis
}
\keyword{datasets}
\keyword{plot}
