% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binmeans.R
\name{binmeans}
\alias{binmeans}
\alias{binmeans.default}
\alias{binmeans.formula}
\title{Compute the mean of a variable for bins of another variable}
\usage{
binmeans(x, y, width = NULL, center = NULL, g = NULL, ...)

\method{binmeans}{default}(x, y, width = NULL, center = NULL, g = NULL, ..., name_g = "colour")

\method{binmeans}{formula}(x, y, width = NULL, center = NULL, ...)
}
\arguments{
\item{x, y}{either two numeric vector for the default method, or a
formula and a data for the formula method}

\item{width}{the width of the bins}

\item{center}{the cuting point of the forcing variable}

\item{g}{a grouping variable}

\item{...}{further arguments}

\item{name_g}{internally used by the geom}
}
\description{
In regression discontinuity, it is usual to plot average values of
the outcome for beans of the forcing variable.
}
