% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg.R
\name{ordreg}
\alias{ordreg}
\title{Ordered regression}
\usage{
ordreg(
  formula,
  data,
  weights,
  subset,
  na.action,
  offset,
  link = c("probit", "logit"),
  start = NULL,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model}

\item{data}{a data frame}

\item{subset, weights, na.action, offset}{see \code{lm}}

\item{link}{one of \code{probit} and \code{logit}}

\item{start}{a vector of starting values, in this case, no
estimation}

\item{...}{further arguments}
}
\value{
an object of class \code{micsr}, see \code{micsr::micsr} for further
details
}
\description{
Maximum-likelihood estimation of a model for which the response is
ordinal
}
\examples{
mod1 <- ordreg(factor(dindx) ~ rhs1 + catchup, fin_reform, link = "logit")
}
\keyword{models}
