% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg.R
\name{ordreg}
\alias{ordreg}
\title{Ordered regression}
\usage{
ordreg(
  formula,
  data,
  weights,
  subset,
  na.action,
  offset,
  link = c("probit", "logit"),
  start = NULL,
  ...
)
}
\arguments{
\item{formula}{a symbolic description of the model, (for the count
component and for the selection equation).}

\item{data}{a data frame,}

\item{subset, weights, na.action, offset}{see \code{stats::lm},}

\item{link}{one of \code{probit} and \code{logit}}

\item{start}{a vector of starting values, in this case, no
estimation}

\item{...}{further arguments}
}
\value{
an object of class \code{micsr}, see \code{micsr::micsr} for further
details
}
\description{
A unified interface perform binomial regression calling \code{lm} to fit
the linear-probability model and \code{glm} to fit the probit and the
logit model.
}
\examples{
mod1 <- ordreg(factor(dindx) ~ rhs1 + catchup, fin_reform, link = "logit")
}
