% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordreg.R
\name{ordreg}
\alias{ordreg}
\alias{fitted.ordreg}
\title{Ordered regression}
\usage{
ordreg(
  formula,
  data,
  weights,
  subset,
  na.action,
  offset,
  contrasts = NULL,
  link = c("probit", "logit", "cloglog"),
  method = c("bfgs", "nr"),
  start = NULL,
  check_gradient = FALSE,
  ...
)

\method{fitted}{ordreg}(object, ..., type = c("outcome", "probabilities"))
}
\arguments{
\item{formula}{a symbolic description of the model}

\item{data}{a data frame}

\item{subset, weights, na.action, offset, contrasts}{see \code{lm}}

\item{link}{one of \code{probit} and \code{logit}}

\item{method}{optimization method}

\item{start}{a vector of starting values, in this case, no
estimation}

\item{check_gradient}{if \code{TRUE} the numeric gradient and hessian
are computed and compared to the analytical gradient and
hessian}

\item{...}{further arguments}

\item{object}{a \code{ordreg} object}

\item{type}{one of \code{"outcome"} or \code{"probabilities"} for the
\code{fitted} method}
}
\value{
an object of class \code{micsr}, see \code{micsr::micsr} for further
details.
}
\description{
Maximum-likelihood estimation of a model for which the response is
ordinal
}
\examples{
mod1 <- ordreg(factor(dindx) ~ rhs1 + catchup, fin_reform, link = "logit")
library(survival)
ud <- unemp_duration \%>\%
      mutate(years = floor(duration / 365),
             years = ifelse(years == 6, 5, years))
mod2 <- ordreg(Surv(years, censored == "no") ~ gender + age + log(1 + wage), ud,
               link = "cloglog", method = "bfgs")
}
\keyword{models}
