% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pbnorm}
\alias{pbnorm}
\title{Bivariate normal probability}
\usage{
pbnorm(z1, z2, rho)
}
\arguments{
\item{z1, z2}{two vectors of normal variates}

\item{rho}{the coefficient of correlation}
}
\value{
the probability
}
\description{
Bivariate normal probability
}
