\name{prod_canada}
\docType{data}
\alias{prod_canada}
\title{Canadian Manufacturing}
\description{a time series of 25 yearly observations from 1946 to 1970

    \emph{country} :  Canada 

    \emph{package} : production

}
\usage{data(prod_canada)}
\format{A dataframe containing:
  \describe{
    \item{year}{year}
    \item{po}{output price}
    \item{qo}{output volume}
    \item{pm}{price index for materials}
    \item{qm}{quantity of materials}
    \item{pl}{price index of productive labor}
    \item{ql}{quantity of productive labor}
    \item{pn}{price index of non-productive labor}
    \item{qn}{quantity of non-productive labor}
    \item{k}{capital stock}
  }
}
\source{original paper, p.375}
\references{
Woodland, A. D. (1977) \dQuote{Estimation of a Variable Profit and of Planning Price Functions for Canadian Manufacturing, 1947-70}, \emph{The Canadian Journal of Economics}, \bold{10(3)}, 355--377.

}
\keyword{datasets}
