% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expreg.R
\name{sargantest}
\alias{sargantest}
\title{Sargan test for GMM models}
\usage{
sargantest(object, ...)
}
\arguments{
\item{object}{a model fitted by GMM}

\item{...}{further arguments}
}
\value{
an \code{htest} object
}
\description{
When a GMM model is over-identified, the set of all the empirical
moment conditions can't be exactly 0. The test of the validity of
the instruments is based on a quadratic form of the vector of the
empirical moments
}
\examples{
cigmales <- cigmales \%>\%
       mutate(age2 = age ^ 2, educ2 = educ ^ 2,
              age3 = age ^ 3, educ3 = educ ^ 3,
              educage = educ * age)
gmm_cig <- expreg(cigarettes ~ habit + price + restaurant + income + age + age2 +
                 educ + educ2 + famsize + race | . - habit + age3 + educ3 +
                 educage + lagprice + reslgth, data = cigmales,
                 twosteps = FALSE)
sargantest(gmm_cig)
}
