% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvuong.R
\docType{data}
\name{turnout}
\alias{turnout}
\title{Turnout}
\format{
a list of three fitted models:
\itemize{
\item group: the group-rule-utilitarian model
\item intens: the intensity model
\item sur: the reduced form SUR model
}
}
\source{
\href{https://www.aeaweb.org/aer/}{American Economic Association data archive}.
}
\description{
these three models are replication in R of stata's
code available on the web site of the American Economic
Association. The estimation is complicated by the fact that
some linear constraints are imposed.
}
\details{
Turnout in Texas liquor referenda
}
\examples{
ndvuong(turnout$group, turnout$intens)
ndvuong(turnout$group, turnout$sur)
ndvuong(turnout$intens, turnout$sur)
}
\references{
\insertRef{COAT:CONL:04}{micsr}
}
\keyword{dataset}
